Reformat all according to coding style

+++ src/utils.c
@@ -5,6 +5,7 @@
 ** section.c
 */
 
+#include <stddef.h>
 #include <elf.h>
 
 Elf64_Shdr *get_section_header(const Elf64_Ehdr *hdr)
@@ -12,12 +13,12 @@ Elf64_Shdr *get_section_header(const Elf64_Ehdr *hdr)
     return ((Elf64_Shdr *) ((void *) hdr + hdr->e_shoff));
 }
 
-Elf64_Shdr *get_section(const Elf64_Ehdr *hdr, int idx)
+Elf64_Shdr *get_section(const Elf64_Ehdr *hdr, size_t idx)
 {
     return (&get_section_header(hdr)[idx]);
 }
 
-Elf64_Sym *get_symbol(const Elf64_Ehdr *hdr, int idx)
+Elf64_Sym *get_symbol(const Elf64_Ehdr *hdr, size_t idx)
 {
     return (&((Elf64_Sym *) hdr)[idx]);
 }
